chatbot_service/requirements.txt
strands-agents>=1.0.0
strands-agents-tools>=0.2.0
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
pydantic>=2.5.0
python-multipart>=0.0.6